Add a Boostrap launcher without dependencies to initialize the system
The main method previously initialized the logger before setting the
system properties, leading to issues with the ForkJoinPool.
